deepseek_qq_bot_istoreos/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
filelock==3.15.4
